spark_job_processor/requirements.txt
#
# This file is autogenerated by pip-compile with Python 3.10
# by the following command:
#
#    pip-compile --resolver=backtracking requirements.in
#
click==8.1.3
    # via flask
flask==2.2.2
    # via
    #   -r requirements.in
    #   flask-sqlalchemy
flask-sqlalchemy==3.0.2
    # via -r requirements.in
greenlet==2.0.1
    # via sqlalchemy
itsdangerous==2.1.2
    # via flask
jinja2==3.1.2
    # via flask
kafka-python==2.0.2
    # via -r requirements.in
markupsafe==2.1.2
    # via
    #   jinja2
    #   werkzeug
psycopg2-binary==2.9.5
    # via -r requirements.in
sqlalchemy==1.4.46
    # via flask-sqlalchemy
werkzeug==2.2.2
    # via flask
